Rename `setTableData` to `getNewStateOnDataChange` and split to it's own file
