+++ src/functions/index.ts
@@ -1,9 +1,14 @@
 export { buildMap } from './build-map'
 export { cloneDeep } from './clone-deep'
-export { getCollatorComparator } from './get-collator-comparator'
+// export { getCollatorComparator } from './get-collator-comparator'
 export { load } from './load'
 export { save } from './save'
-export { transformData } from './transform-data'
-export { sortCompare } from './sort-compare'
+// export { transformData } from './transform-data'
+// export { sortCompare } from './sort-compare'
 export { warnDeprecated } from './warn-deprecated'
 export { warnInfo } from './warn-info'
+export {
+    getDisplayData,
+    getNewStateOnDataChange,
+    sortTable
+} from './get-new-state-on-data-change'
